setup contact plans for 3 scenarios. see details:
I learned that RSSI noise affects contact plan generation a lot. I used 0 noise for contact plan generation, but when we run the sims we should have a bit of noise. I think this is fine (actual contacts may vary slightly from contact plan).

Right now we're essentially being opportunistic (if we see an unexpected connection not in the contact plan, we use it regardless). The contact plan is still useful for routing next hops across the network. The only difference is that we might see the next hop earlier or later than planned.